fix: swagger

+++ src/main/java/main/java/insper/store/products/ProductService.java
@@ -27,7 +27,11 @@ public class ProductService {
     }
 
     public void delete(@NonNull String id) {
-        productRepository.deleteById(id);
+        if(productRepository.existsById(id)){
+            productRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Product not found");
+        }
     }
 
     public List<ProductModel> findAll() {
